Add script for time of matrix multiplication
I wrote this script recently for another class. It multiplies NxN matrices together and times how long it takes to compute the multiplication. It does that for matrices of various sizes, plots the calculation time versus matrix size and uses scipy.optimize to find how computation time scales with size. It also does this for three different methods of multiplication: two built-in python methods and one hand written method. 

This could be easily adapted to include Einstein notation and extended to other matrix operations. It needs some documentation to describe what's going on.